requirements.txt
pynput>=1.7.6
PyQt6>=6.4.0
pystray>=0.19.0
Pillow>=10.0.0
keyboard>=0.13.5